Change type of Compensation Value to float

+++ src/pages/Offers/OfferTable.js
@@ -27,7 +27,7 @@ export const columns = [
     id: "wage_type",
     type: ["Salary", "Hourly", "On-completion"]
   },
-  { name: "Compensation Value", id: "wage_value", type: "int" }
+  { name: "Compensation Value", id: "wage_value", type: "float" }
 ];
 
 const OfferTable = ({ offers }) => {
